BackEnd/requirements.txt
fastapi
uvicorn
python-dotenv
openai
gtts
requests
speechrecognition
